blindfold/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blindfold"
version = "0.0.0"
description = "A tool for blind mode operations"
readme = "README.md"
authors = [{name = "Author", email = "author@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]

[project.scripts]
blindfold = "blindfold.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
blindfold = ["data/**/*.yaml"]